updated nonlinear optimization routines
